age -> birthday

+++ src/main/java/com/isep/project/repository/AuthRepository.java
@@ -4,7 +4,7 @@ import com.isep.project.entity.Auth;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AuthRepository extends JpaRepository<Auth, Integer>, JpaSpecificationExecutor<Auth>
+public interface AuthRepository extends JpaRepository<Auth, Long>, JpaSpecificationExecutor<Auth>
 {
 
 }
