+++ src/main/java/com/isep/project/repository/UserRepository.java
@@ -4,7 +4,7 @@ import com.isep.project.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User>
+public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User>
 {
 
 }
